# [TRITON] Optimize per token fp8 quant kernel