--- a/aiter/ops/triton/quant.py
+++ b/aiter/ops/triton/quant.py
@@ -4,37 +4,11 @@ import torch
 
 
 @triton.jit
-def _per_tensor_fp8_quant(x_in: torch.Tensor, 
-                                scale_in: torch.Tensor
-): 
-    """
-    Quantize to fp8 with the scale provided
-
-    Parameters:
-    x_in: Tensor to be quantized
-    scale_in: Scale value of shape [1]. Used for all the elements in the x_in
-
-    Returns:
-    qx: quantized values. Must have same shape as x_in
-
-    Note: x_in, scale_in must be loaded in the caller triton jitted function
-    """
-    qx = x_in / scale_in
-    qx = qx.to(tl.float8e4b8)
-
-    return qx
-
-@triton.jit
 def _static_per_tensor_fp8_quant_kernel(qx_ptr: torch.Tensor, 
                                         x_in_ptr: torch.Tensor, 
                                         scale_in_ptr: torch.Tensor,
-                                        rows: int,
                                         cols: int,
-                                        qx_stride_r: int,
-                                        qx_stride_c: int,
                                         x_in_stride_r: int,
-                                        x_in_stride_c: int,
-                                        BLOCK_SIZE: tl.constexpr,
                                         NUM_COL_POW2: tl.constexpr
 ):
     """
@@ -42,14 +16,17 @@ def _static_per_tensor_fp8_quant_kernel(qx_ptr: torch.Tensor,
     """
 
     pid = tl.program_id(axis=0)
+    tl.assume(pid > 0)
+    tl.assume(x_in_stride_r > 0)
 
-    offs = pid*BLOCK_SIZE*qx_stride_r + tl.arange(0, BLOCK_SIZE)[:, None]*qx_stride_r + tl.arange(0, NUM_COL_POW2)[None, :]
-    mask = (tl.arange(0, BLOCK_SIZE) < rows)[:, None] &  (tl.arange(0, NUM_COL_POW2) < cols)[None, :]   
-    x = tl.load(x_in_ptr + offs, mask = mask)
+    offs = pid * x_in_stride_r + tl.arange(0, NUM_COL_POW2)
+    mask = tl.arange(0, NUM_COL_POW2) < cols
+    x = tl.load(x_in_ptr + offs, mask = mask, cache_modifier=".cg")
 
     scale = tl.load(scale_in_ptr)
+    scale_recip = 1 / scale
 
-    qx = _per_tensor_fp8_quant(x, scale)
+    qx = (x * scale_recip).to(qx_ptr.dtype.element_ty)
 
     tl.store(qx_ptr + offs, qx, mask=mask)
 
@@ -65,19 +42,13 @@ def static_per_tensor_fp8_quant(qx: torch.Tensor,
     assert scale_in.numel() == 1 #only single scale value 
     rows = x_in.shape[0]
     cols = x_in.shape[1]
-    BLOCK_SIZE = 32
     NUM_COL_POW2 = triton.next_power_of_2(cols)
-    grid = lambda meta: (triton.cdiv(rows,meta['BLOCK_SIZE']),)  
+    grid = lambda meta: (rows,)
     _static_per_tensor_fp8_quant_kernel[grid](qx, 
                                             x_in, 
                                             scale_in, 
-                                            rows, 
                                             cols, 
-                                            qx.stride(0), 
-                                            qx.stride(1), 
                                             x_in.stride(0),
-                                            x_in.stride(1),
-                                            BLOCK_SIZE=BLOCK_SIZE,
                                             NUM_COL_POW2=NUM_COL_POW2
                                             )
 
@@ -85,17 +56,10 @@ def static_per_tensor_fp8_quant(qx: torch.Tensor,
 
 
 @triton.jit
-def _dynamic_per_tensor_fp8_quant_kernel(qx_ptr: torch.Tensor, 
+def _dynamic_per_tensor_fp8_quant_kernel(x_in_ptr: torch.Tensor,
                                         scale_out_ptr: torch.Tensor,
-                                        x_max_ptr: torch.Tensor,
-                                        x_in_ptr: torch.Tensor, 
-                                        rows: int,
                                         cols: int,
-                                        qx_stride_r: int,
-                                        qx_stride_c: int,
                                         x_in_stride_r: int,
-                                        x_in_stride_c: int,
-                                        BLOCK_SIZE: tl.constexpr,
                                         NUM_COL_POW2: tl.constexpr,
                                         FP8_MAX: tl.constexpr
 ):
@@ -104,22 +68,16 @@ def _dynamic_per_tensor_fp8_quant_kernel(qx_ptr: torch.Tensor,
     """
 
     pid = tl.program_id(axis=0)
+    tl.assume(pid > 0)
+    tl.assume(x_in_stride_r > 0)
 
-    offs = pid*BLOCK_SIZE*x_in_stride_r + tl.arange(0, BLOCK_SIZE)[:, None]*x_in_stride_r + tl.arange(0, NUM_COL_POW2)[None, :]
-    mask = (pid*BLOCK_SIZE + tl.arange(0, BLOCK_SIZE) < rows)[:, None] &  (tl.arange(0, NUM_COL_POW2) < cols)[None, :]   
-    x = tl.load(x_in_ptr + offs, mask = mask)
+    offs = pid * x_in_stride_r + tl.arange(0, NUM_COL_POW2)
+    mask = tl.arange(0, NUM_COL_POW2) < cols
+    x = tl.load(x_in_ptr + offs, mask = mask, cache_modifier=".cg")
 
     m = tl.max(tl.abs(x))
-    tl.atomic_max(x_max_ptr, m)
-    x_max = tl.load(x_max_ptr)
+    tl.atomic_max(scale_out_ptr, m / FP8_MAX, sem="relaxed")
     
-    scale_out = x_max / FP8_MAX
-    
-    qx = _per_tensor_fp8_quant(x, scale_out)
-
-    tl.store(qx_ptr + offs, qx, mask=mask)
-    tl.store(scale_out_ptr, scale_out)
-
 
 def dynamic_per_tensor_fp8_quant(qx: torch.Tensor, 
                                 x_in: torch.Tensor,
@@ -131,63 +89,58 @@ def dynamic_per_tensor_fp8_quant(qx: torch.Tensor,
     """
     rows = x_in.shape[0]
     cols = x_in.shape[1]
-    x_max = torch.zeros((1), dtype=torch.float32, device='cuda')
-    BLOCK_SIZE = 32
     NUM_COL_POW2 = triton.next_power_of_2(cols)
-    grid = lambda meta: (triton.cdiv(rows,meta['BLOCK_SIZE']),)  
-    _dynamic_per_tensor_fp8_quant_kernel[grid](qx, 
+    grid = lambda meta: (rows,)
+    _dynamic_per_tensor_fp8_quant_kernel[grid](x_in,
                                             scale_out,
-                                            x_max,
-                                            x_in, 
-                                            rows, 
-                                            cols, 
-                                            qx.stride(0), 
-                                            qx.stride(1), 
+                                            cols,
                                             x_in.stride(0),
-                                            x_in.stride(1),
-                                            BLOCK_SIZE=BLOCK_SIZE,
                                             NUM_COL_POW2=NUM_COL_POW2,
-                                            FP8_MAX=torch.finfo(torch.float8_e4m3fnuz).max
+                                            FP8_MAX=torch.finfo(qx.dtype).max
                                             )
 
+    _static_per_tensor_fp8_quant_kernel[grid](qx,
+                                            x_in,
+                                            scale_out,
+                                            cols,
+                                            x_in.stride(0),
+                                            NUM_COL_POW2=NUM_COL_POW2)
+
     return qx, scale_out
+
 
 @triton.jit
 def _dynamic_per_token_fp8_quant_kernel(qx_ptr: torch.Tensor,
                                         scale_out_ptr: torch.Tensor,
                                         x_in_ptr: torch.Tensor,
-                                        rows: int,
                                         cols: int,
-                                        qx_stride_r: int,
-                                        qx_stride_c: int,
                                         x_in_stride_r: int,
-                                        x_in_stride_c: int,
-                                        BLOCK_SIZE: tl.constexpr,
                                         NUM_COL_POW2: tl.constexpr,
                                         FP8_MAX: tl.constexpr,
-                                        QDTYPE: tl.constexpr
 ):
     """
     #TODO: Add Doc
     """
 
     pid = tl.program_id(axis=0)
+    tl.assume(pid > 0)
+    tl.assume(x_in_stride_r > 0)
 
-    offs = pid*BLOCK_SIZE*x_in_stride_r + tl.arange(0, BLOCK_SIZE)[:, None]*x_in_stride_r + tl.arange(0, NUM_COL_POW2)[None, :]
-    mask = (pid*BLOCK_SIZE + tl.arange(0, BLOCK_SIZE) < rows)[:, None] &  (tl.arange(0, NUM_COL_POW2) < cols)[None, :]   
-    x = tl.load(x_in_ptr + offs, mask = mask)
+    offs = pid * x_in_stride_r + tl.arange(0, NUM_COL_POW2)
+    mask = tl.arange(0, NUM_COL_POW2) < cols
+    x = tl.load(x_in_ptr + offs, mask=mask, cache_modifier=".cg")
 
     m = tl.max(tl.abs(x), axis=-1)
     scale_out = m / FP8_MAX
+    scale_recip = 1 / scale_out
 
-    qx = x / scale_out[:, None]
-    qx = qx.to(QDTYPE)
+    qx = x * scale_recip
+    qx = qx.to(qx_ptr.dtype.element_ty)
 
-    tl.store(qx_ptr + offs, qx, mask=mask)
+    scale_offs = pid
+    tl.store(scale_out_ptr + scale_offs, scale_out)
 
-    scale_offs = pid*BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
-    scale_mask = pid*BLOCK_SIZE + tl.arange(0, BLOCK_SIZE) < rows
-    tl.store(scale_out_ptr + scale_offs, scale_out, mask=scale_mask)
+    tl.store(qx_ptr + offs, qx, mask=mask, cache_modifier=".cs")
 
 
 def dynamic_per_token_fp8_quant(qx: torch.Tensor,
@@ -202,31 +155,15 @@ def dynamic_per_token_fp8_quant(qx: torch.Tensor,
     """
     rows = x_in.shape[0]
     cols = x_in.shape[1]
-    BLOCK_SIZE = 32
     NUM_COL_POW2 = triton.next_power_of_2(cols)
-    grid = lambda meta: (triton.cdiv(rows,meta['BLOCK_SIZE']),)  
-    qdtype = torchDtype2Triton(quant_dtype)
+    grid = lambda meta: (rows,)
     _dynamic_per_token_fp8_quant_kernel[grid](qx, 
                                             scale_out,
                                             x_in, 
-                                            rows, 
                                             cols, 
-                                            qx.stride(0), 
-                                            qx.stride(1), 
                                             x_in.stride(0),
-                                            x_in.stride(1),
-                                            BLOCK_SIZE=BLOCK_SIZE,
                                             NUM_COL_POW2=NUM_COL_POW2,
                                             FP8_MAX=dtypeMax,
-                                            QDTYPE=qdtype
                                             )
 
     return qx, scale_out
-
-
-def torchDtype2Triton(dtype):
-    tmp = {
-        torch.int8: tl.int8,
-        torch.float8_e4m3fnuz: tl.float8e4b8,
-    }
-    return tmp[dtype]