1) Made the grid equal to num rows. Previously it was one WG every 32 rows. (60us -> 5.5us)
2) Added cache modifiers for ld/st (5.5us -> 4.7us) 
3) Explicit reciprocal to avoid Newton Raphson steps on result (4.7 -> 3.5us)

Results on 128 x 8192 tensor. Roofline ~1us.

This is now exposing HBM latency. Can't really optimize more without significant effort.